Bug 571190 build the solution on Framework services host
Extract miner
