+++ bundles/org.eclipse.passage.lbc.base/src/org/eclipse/passage/lbc/internal/base/mine/Conditions.java
@@ -20,17 +20,12 @@ import java.util.stream.Collectors;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.eclipse.passage.lbc.internal.base.api.RawRequest;
-import org.eclipse.passage.lic.floating.FloatingFileExtensions;
-import org.eclipse.passage.lic.internal.api.LicensedProduct;
 import org.eclipse.passage.lic.internal.api.PassageAction;
 import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
-import org.eclipse.passage.lic.internal.api.conditions.ConditionMiningTarget;
 import org.eclipse.passage.lic.internal.api.conditions.ConditionPack;
-import org.eclipse.passage.lic.internal.base.conditions.mining.LocalConditions;
 import org.eclipse.passage.lic.internal.base.diagnostic.DiagnosticExplained;
 import org.eclipse.passage.lic.internal.base.diagnostic.NoErrors;
 import org.eclipse.passage.lic.internal.base.io.LicensingFolder;
-import org.eclipse.passage.lic.internal.base.io.PathFromLicensedProduct;
 import org.eclipse.passage.lic.internal.base.io.UserHomePath;
 import org.eclipse.passage.lic.internal.net.api.handle.NetResponse;
 import org.eclipse.passage.lic.internal.net.handle.EObjectTransfer;
@@ -57,7 +52,7 @@ public final class Conditions implements Supplier<NetResponse> {
 	public NetResponse get() {
 		log.debug(String.format("Mining conditions for product %s", data.product().get())); //$NON-NLS-1$
 		ServiceInvocationResult<Collection<ConditionPack>> conditions = //
-				new Miner(source, data.user().get())//
+				new FloatingConditions(source, data.user().get())//
 						.all(data.product().get());
 		if (!conditions.data().isPresent()) {
 			return new Failure.OperationFailed(//
@@ -78,25 +73,6 @@ public final class Conditions implements Supplier<NetResponse> {
 						.flatMap(pack -> pack.conditions().stream())//
 						.collect(Collectors.toList())//
 		).get();
-	}
-
-	private final class Miner extends LocalConditions {
-
-		private final Supplier<Path> base;
-
-		Miner(Supplier<Path> base, String user) {
-			super(//
-					new ConditionMiningTarget.Local().child("floating-server"), //$NON-NLS-1$
-					new ReassemblingMiningEquipment(user, base), //
-					new FloatingFileExtensions.FloatingLicenseEncrypted());
-			this.base = base;
-		}
-
-		@Override
-		protected Supplier<Path> base(LicensedProduct product) {
-			return new PathFromLicensedProduct(base, product);
-		}
-
 	}
 
 }
