+++ bundles/org.eclipse.passage.lbc.base/src/org/eclipse/passage/lbc/internal/base/mine/FloatingConditions.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -15,25 +15,27 @@ package org.eclipse.passage.lbc.internal.base.mine;
 import java.nio.file.Path;
 import java.util.function.Supplier;
 
+import org.eclipse.passage.lic.floating.FloatingFileExtensions;
 import org.eclipse.passage.lic.internal.api.LicensedProduct;
-import org.eclipse.passage.lic.internal.api.LicensingException;
 import org.eclipse.passage.lic.internal.api.conditions.ConditionMiningTarget;
-import org.eclipse.passage.lic.internal.api.conditions.mining.MiningEquipment;
-import org.eclipse.passage.lic.internal.api.conditions.mining.MiningTool;
+import org.eclipse.passage.lic.internal.base.conditions.mining.LocalConditions;
+import org.eclipse.passage.lic.internal.base.io.PathFromLicensedProduct;
 
-public final class ReassemblingMiningEquipment implements MiningEquipment {
+final class FloatingConditions extends LocalConditions {
 
-	private final String user;
 	private final Supplier<Path> base;
 
-	ReassemblingMiningEquipment(String user, Supplier<Path> base) {
-		this.user = user;
+	FloatingConditions(Supplier<Path> base, String user) {
+		super(//
+				new ConditionMiningTarget.Local().child("floating-server"), //$NON-NLS-1$
+				new ReassemblingMiningEquipment(user, base), //
+				new FloatingFileExtensions.FloatingLicenseEncrypted());
 		this.base = base;
 	}
 
 	@Override
-	public MiningTool tool(LicensedProduct product, ConditionMiningTarget miner) throws LicensingException {
-		return new ReassemblingMiningTool(product, user, base, miner);
+	protected Supplier<Path> base(LicensedProduct product) {
+		return new PathFromLicensedProduct(base, product);
 	}
 
 }
