+++ bundles/org.eclipse.passage.lic.api/META-INF/MANIFEST.MF
@@ -8,37 +8,38 @@ Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.eclipse.passage.lic.internal.api;
-  x-friends:="org.eclipse.passage.lbc.api,
+  x-friends:="org.eclipse.passage.lac,
+   org.eclipse.passage.lbc.api,
    org.eclipse.passage.lbc.base,
-   org.eclipse.passage.loc.workbench,
-   org.eclipse.passage.loc.dashboard.ui,
-   org.eclipse.passage.loc.features.ui,
-   org.eclipse.passage.loc.licenses.ui,
-   org.eclipse.passage.loc.products.ui,
-   org.eclipse.passage.loc.users.ui,
-   org.eclipse.passage.loc.api,
-   org.eclipse.passage.lic.base,
-   org.eclipse.passage.lic.jface,
-   org.eclipse.passage.lic.hc,
-   org.eclipse.passage.lic.equinox,
-   org.eclipse.passage.lic.equinox.tests,
-   org.eclipse.passage.lic.base.tests,
-   org.eclipse.passage.lic.api.tests,
    org.eclipse.passage.lbc.base.tests,
-   org.eclipse.passage.loc.licenses.core,
-   org.eclipse.passage.lic.net,
-   org.eclipse.passage.lic.emf,
-   org.eclipse.passage.seal.demo.tests,
-   org.eclipse.passage.lic.hc.tests,
-   org.eclipse.passage.seal.demo,
+   org.eclipse.passage.lbc.seal,
+   org.eclipse.passage.lic.api.tests,
+   org.eclipse.passage.lic.base,
+   org.eclipse.passage.lic.base.tests,
    org.eclipse.passage.lic.bc,
    org.eclipse.passage.lic.e4.ui,
+   org.eclipse.passage.lic.emf,
+   org.eclipse.passage.lic.equinox,
+   org.eclipse.passage.lic.equinox.tests,
+   org.eclipse.passage.lic.hc,
+   org.eclipse.passage.lic.hc.tests,
+   org.eclipse.passage.lic.jface,
    org.eclipse.passage.lic.json,
-   org.eclipse.passage.lic.oshi,
-   org.eclipse.passage.loc.products.core,
-   org.eclipse.passage.lac,
    org.eclipse.passage.lic.licenses.migration,
-   org.eclipse.passage.lac.seal",
+   org.eclipse.passage.lic.net,
+   org.eclipse.passage.lic.oshi,
+   org.eclipse.passage.loc.api,
+   org.eclipse.passage.loc.dashboard.ui,
+   org.eclipse.passage.loc.features.ui,
+   org.eclipse.passage.loc.licenses.core,
+   org.eclipse.passage.loc.licenses.ui,
+   org.eclipse.passage.loc.products.core,
+   org.eclipse.passage.loc.products.ui,
+   org.eclipse.passage.loc.users.ui,
+   org.eclipse.passage.loc.workbench,
+   org.eclipse.passage.seal.demo,
+   org.eclipse.passage.seal.demo.tests,
+   org.eclipse.passage.lac.gear",
  org.eclipse.passage.lic.internal.api.access;
   x-friends:="org.eclipse.passage.lic.base,
    org.eclipse.passage.lic.e4.ui,
